[skip ci] fix test and move the test server to internal
