+++ internal/common.go
@@ -1,4 +1,4 @@
-package runner
+package internal
 
 import (
 	"net"
@@ -6,14 +6,19 @@ import (
 
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/credentials"
+	"google.golang.org/grpc/reflection"
 
 	"github.com/bojand/ghz/internal/helloworld"
 )
 
-var port string
-var localhost string
+// TestPort is the port
+var TestPort string
 
-func startServer(secure bool) (*helloworld.Greeter, *grpc.Server, error) {
+// TestLocalhost is the localhost
+var TestLocalhost string
+
+// StartServer starts server
+func StartServer(secure bool) (*helloworld.Greeter, *grpc.Server, error) {
 	lis, err := net.Listen("tcp", ":0")
 	if err != nil {
 		return nil, nil, err
@@ -33,10 +38,10 @@ func startServer(secure bool) (*helloworld.Greeter, *grpc.Server, error) {
 
 	gs := helloworld.NewGreeter()
 	helloworld.RegisterGreeterServer(s, gs)
-	// reflection.Register(s)
+	reflection.Register(s)
 
-	port = strconv.Itoa(lis.Addr().(*net.TCPAddr).Port)
-	localhost = "localhost:" + port
+	TestPort = strconv.Itoa(lis.Addr().(*net.TCPAddr).Port)
+	TestLocalhost = "localhost:" + TestPort
 
 	go func() {
 		s.Serve(lis)
